+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 Release Notes
 =============
 
-## Next Release
+## 0.101.0
+
+Released January 20, 2017
 
 **New**